Make DataScript static and align runtime performance

+++ runtime/core_set.ml
@@ -1,4 +1,4 @@
-module Int_set = Set.Make (Int64)
+module Int_set = Set.Make (Int)
 module Float_set = Set.Make (Float)
 module String_set = Set.Make (String)
 
@@ -11,7 +11,7 @@ end
 module Bool_set = Set.Make (Bool_order)
 
 module Int_list_set = Set.Make (struct
-  type t = int64 list
+  type t = int list
 
   let compare = Stdlib.compare
 end)
@@ -35,7 +35,7 @@ module Bool_list_set = Set.Make (struct
 end)
 
 module Int_vector_set = Set.Make (struct
-  type t = int64 Rrbvec.t
+  type t = int Rrbvec.t
 
   let compare = Stdlib.compare
 end)
@@ -95,7 +95,7 @@ module Dynamic_vector_vector_set = Set.Make (struct
 end)
 
 module Int_vector_vector_set = Set.Make (struct
-  type t = int64 Rrbvec.t Rrbvec.t
+  type t = int Rrbvec.t Rrbvec.t
 
   let compare = Stdlib.compare
 end)
